IOTCORE-389 - bring back legacy alignment for Arm C6
Quite a few of the scatter files are not (yet) aligned to 8-byte
boundaries and therefore the removal of legacy alignment feature
(which is under deprecation warning, but it actually not YET
deprecated) broke quite a few builds to this error:

Error: L6244E: Exec region RW_IRAM1 address (0x200001ac) not aligned on a 8 byte boundary.

We must bring this option now back to fix the builds.

This option to ld (--legacyalign) can only be removed once all of
the scatter files have been fixed.