Revert "Turn off MPU on targets failing OOB"
The cause of the pelion-enablement test failures for Nuvoton boards
is not caused by or related to the MPU so this patch re-enables the
MPU for these boards.

The cause of the  pelion-enablement test failures for the
UBLOX_C030_U201 has been identified and corrected so this patch
re-enables the MPU for this board.

This reverts commit d3d40bcbde60fd9a1aa79bc30f9942f61ffeb1e2.